Split agnostic into non-gui and gui
